refactor(client): test directory
