+++ client/src/components/__tests__/LobbyCompontent.js
@@ -1,7 +1,6 @@
+import {fireEvent, render, screen} from '@testing-library/react';
 import React from 'react';
-import {render, screen, fireEvent} from '@testing-library/react';
-
-import Lobby from '../components/LobbyComponent.js';
+import Lobby from '../LobbyComponent.js';
 
 describe('Lobby', () => {
   test('Renders room code', () => {
