+++ client/src/components/__tests__/PresentationComponent.js
@@ -1,6 +1,6 @@
 import {render, screen} from '@testing-library/react';
 import React from 'react';
-import Presentation from '../components/PresentationComponent.js';
+import Presentation from '../PresentationComponent.js';
 
 describe('Presentation', () => {
   test('Renders the', () => {
